requirements.txt
openai==1.3.1
streamlit==1.26.0
python-dotenv==1.0.0
